Added projectile tier selection

+++ Assets/Scripts/Functions/Other/Constant.cs
@@ -24,7 +24,7 @@ public class Constant : Function
 		get { return value.ToString(); }
 	}
 
-	private float value;
+	private int value;
 
 	public Constant() {
 		value = Random.Range(-10, 11);
